chore: Add support for Rust implementation in Advent of Code 2023
This commit adds support for Rust implementation in Advent of Code 2023.
It includes a link to a video tutorial on setting up Rust for Advent of
Code and expresses excitement for the upcoming challenge. This
enhancement enables users to participate in the event using Rust,
providing them with a new language option. Enjoy the coding experience!

References: #AdventOfCode2023, #RustImplementation